% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventDetection.R
\name{.txs2genesym}
\alias{.txs2genesym}
\title{Adding gene symbol annotations}
\usage{
.txs2genesym(txs, unique.genesyms = TRUE)
}
\arguments{
\item{txs}{A list of transcript ids in UCSC format.}

\item{unique.genesyms}{TRUE or FALSE. If TRUE, the converted gene symbols will remove duplicates.}
}
\value{
A list of names in gene symbols
}
\description{
Adding gene symbol annotations
}
\details{
This is an internal function which takes a list of txs in UCSC id format as input 
and convert the txs to gene symbol.
}
